fix: displaying checkbox and refactor fetching streaming payments
